anillo_2021_01_25/opc_controller_pc3/requirements.txt
numpy==1.19.4
pandas==1.1.4
opcua==0.98.12
cassandra-driver==3.24.0
